feat: add theme.state opacity tokens and split theme types
