+++ src/theme/schemes/DarkTheme.tsx
@@ -1,6 +1,7 @@
 import { baseTheme } from './base';
 import { tokens } from '../tokens';
 import { buildScheme } from '../tokens/sys/color/roles';
+import { defaultState } from '../tokens/sys/state';
 import type { Theme } from '../types';
 
 export const DarkTheme: Theme = {
@@ -8,4 +9,5 @@ export const DarkTheme: Theme = {
   dark: true,
   mode: 'adaptive',
   colors: buildScheme(tokens.md.ref.palette, tokens.md.ref, { mode: 'dark' }),
+  state: defaultState,
 };
