+++ src/theme/tokens/sys/state.ts
@@ -1,0 +1,13 @@
+import type { ThemeState } from '../../types';
+import { tokens } from '../index';
+
+const { stateOpacity } = tokens.md.ref;
+
+export const defaultState: ThemeState = {
+  opacity: {
+    hovered: stateOpacity.hovered,
+    focused: stateOpacity.focused,
+    pressed: stateOpacity.pressed,
+    dragged: stateOpacity.dragged,
+  },
+};
