+++ src/theme/types/elevation.ts
@@ -1,0 +1,14 @@
+export type Elevation = 0 | 1 | 2 | 3 | 4 | 5;
+
+export enum ElevationLevels {
+  'level0',
+  'level1',
+  'level2',
+  'level3',
+  'level4',
+  'level5',
+}
+
+export type ElevationColors = {
+  [key in keyof typeof ElevationLevels]: string;
+};
