+++ src/theme/types/state.ts
@@ -1,0 +1,5 @@
+export type InteractionState = 'hovered' | 'focused' | 'pressed' | 'dragged';
+
+export type ThemeState = {
+  opacity: Record<InteractionState, number>;
+};
