+++ src/theme/types/typography.ts
@@ -1,0 +1,61 @@
+export type Font = {
+  fontFamily: string;
+  fontWeight?:
+    | 'normal'
+    | 'bold'
+    | '100'
+    | '200'
+    | '300'
+    | '400'
+    | '500'
+    | '600'
+    | '700'
+    | '800'
+    | '900';
+  fontStyle?: 'normal' | 'italic' | undefined;
+};
+
+export type Fonts = {
+  regular: Font;
+  medium: Font;
+  light: Font;
+  thin: Font;
+};
+
+export enum TypescaleKey {
+  displayLarge = 'displayLarge',
+  displayMedium = 'displayMedium',
+  displaySmall = 'displaySmall',
+
+  headlineLarge = 'headlineLarge',
+  headlineMedium = 'headlineMedium',
+  headlineSmall = 'headlineSmall',
+
+  titleLarge = 'titleLarge',
+  titleMedium = 'titleMedium',
+  titleSmall = 'titleSmall',
+
+  labelLarge = 'labelLarge',
+  labelMedium = 'labelMedium',
+  labelSmall = 'labelSmall',
+
+  bodyLarge = 'bodyLarge',
+  bodyMedium = 'bodyMedium',
+  bodySmall = 'bodySmall',
+}
+
+export type TypescaleStyle = {
+  fontFamily: string;
+  letterSpacing: number;
+  fontWeight: Font['fontWeight'];
+  lineHeight: number;
+  fontSize: number;
+  fontStyle?: Font['fontStyle'];
+};
+
+export type Typescale =
+  | {
+      [key in TypescaleKey]: TypescaleStyle;
+    } & {
+      ['default']: Omit<TypescaleStyle, 'lineHeight' | 'fontSize'>;
+    };
